Finalizado de implementación de creación de asegurados.

+++ backend/InsuranceAPI/Repositories/ProducerRepository.cs
@@ -3,8 +3,8 @@
 namespace InsuranceAPI.Repositories {
     public interface IProducerRepository {
         public List<Producer> getAll();
-        public Producer getById(int id);
-        public List<Producer> getByIds(List<int> ids);
+        public Producer getById(long id);
+        public List<Producer> getByIds(List<long> ids);
     }
 
     public class ProducerRepository : IProducerRepository{
@@ -18,15 +18,15 @@ namespace InsuranceAPI.Repositories {
             return _context.Producers.ToList();
         }
 
-        public Producer getById(int id) {
+        public Producer getById(long id) {
             return (from prod in _context.Producers
                     where prod.Id == id
                     select prod).First();
         }
 
-        public List<Producer> getByIds(List<int> ids) {
+        public List<Producer> getByIds(List<long> ids) {
             List<Producer> ret = new List<Producer>();
-            foreach(int id in ids) 
+            foreach(long id in ids) 
                 ret.Add(getById(id));
             return ret;
         }
